apps/rag-service/requirements.txt
# Core API framework
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv

# PDF processing
pymupdf

# Embeddings + ML
torch
sentence-transformers
faiss-cpu
numpy
tqdm

# Storage / cloud
boto3
requests
